requirements.txt
pyyaml
python-dotenv
anthropic
openai
tiktoken
